backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "saludpe-backend"
version = "0.1.0"
description = "Backend API for SaludPe"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.16.0,<2.0.0",
  "email-validator>=2.2.0,<3.0.0",
  "fastapi[standard]>=0.135.1,<0.136.0",
  "pydantic-settings>=2.10.1,<3.0.0",
  "psycopg[binary]>=3.2.9,<4.0.0",
  "pwdlib[argon2]>=0.2.1,<0.3.0",
  "PyJWT>=2.10.1,<3.0.0",
  "sqlalchemy>=2.0.41,<3.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1,<0.29.0",
  "mypy>=1.16.1,<2.0.0",
  "pytest>=8.4.1,<9.0.0",
  "ruff>=0.12.3,<0.13.0",
]

[tool.fastapi]
entrypoint = "app.main:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.setuptools.packages.find]
where = ["."]
